using abstact handle backend class
